pdf_reader/pdf_extracter/requirements.txt
flask
flask-cors
PyPDF2
pytesseract
pdf2image
Pillow
python-dotenv
langdetect
